resolve: move components to different files
